pyproject.toml
[project]
name = "sibur-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikita_Dzigurda",email = "ghataju@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "response (>=0.5.0,<0.6.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "sibur_project", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
